requirements.txt
requests
Pillow
easyocr
beautifulsoup4
